split out consumer code
